+++ types/index.d.ts
@@ -379,6 +379,7 @@ declare namespace Knex {
     | string
     | number
     | boolean
+    | null
     | Date
     | Array<string>
     | Array<number>
@@ -1366,6 +1367,7 @@ declare namespace Knex {
   interface Raw<TResult = any>
     extends events.EventEmitter,
       ChainableInterface<ResolveResult<TResult>> {
+    timeout(ms: number, options?: {cancel?: boolean}): Raw<TResult>;
     wrap<TResult2 = TResult>(before: string, after: string): Raw<TResult>;
     toSQL(): Sql;
     queryContext(context: any): Raw<TResult>;
@@ -1378,7 +1380,24 @@ declare namespace Knex {
     <TResult2 = TResult>(sql: string, bindings: readonly RawBinding[] | ValueDict): Raw<TResult2>;
   }
 
+  const RefMemberTag: unique symbol;
+
   interface Ref<TSrc extends string, TMapping extends {}> extends Raw<string> {
+      // TypeScript can behave weirdly if type parameters are not
+      // actually used in the members of type.
+      //
+      // See: https://github.com/knex/knex/issues/3932
+      //
+      // We simply need to propagate the type context so that we can extract
+      // them later, but we just add a "phantom" property so that typescript
+      // doesn't think that these parameters are unused
+      //
+      // Because unique symbol is used here, there is no way to actually
+      // access this at runtime
+      [RefMemberTag]: {
+        src: TSrc,
+        mapping: TMapping
+      };
       withSchema(schema: string): this;
       as<TAlias extends string>(alias: TAlias): Ref<TSrc, {[K in TAlias]: TSrc}>;
   }
@@ -1546,6 +1565,11 @@ declare namespace Knex {
     text(columnName: string, textType?: string): ColumnBuilder;
     string(columnName: string, length?: number): ColumnBuilder;
     float(
+      columnName: string,
+      precision?: number,
+      scale?: number
+    ): ColumnBuilder;
+    double(
       columnName: string,
       precision?: number,
       scale?: number
@@ -1941,9 +1965,11 @@ declare namespace Knex {
 
   interface SeederConfig {
     extension?: string;
-    directory?: string;
+    directory?: string | string[];
     loadExtensions?: readonly string[];
     specific?: string;
+    recursive?: boolean;
+    sortDirsSeparately?: boolean;
   }
 
   class Seeder {